Learn about eventHandling and DOM Manipulation
Learned about transition( is used to define how to do it i.e with what ease, how much time and which css property is being changed) , transform(used to change the shape and size of element)

Introduced to audio element and learned how to add class to an element, how add event listeners. learned about keydown, keypress and transitionend events.